requirements.txt
Flask==2.1.2
line-bot-sdk==2.2.1
gunicorn==20.1.0
requests==2.26.0
psycopg2